Revert "Merge branch 'ben-branch'"
This reverts commit 4e9d5adee573a9aceae3b0c389fed9495b67a8d4

+++ sales/api/sales_rest/views.py
@@ -2,7 +2,6 @@ from common.json import ModelEncoder
 from django.http import JsonResponse
 from django.views.decorators.http import require_http_methods
 import json
-from django.shortcuts import render
 
 from .models import AutomobileVO, Salesperson, Sale, Customer
 
@@ -10,6 +9,7 @@ from .models import AutomobileVO, Salesperson, Sale, Customer
 class AutomobileVOEncoder(ModelEncoder):
     model = AutomobileVO
     properties = [
+        "id",
         "vin",
         "sold",
     ]
@@ -18,21 +18,21 @@ class AutomobileVOEncoder(ModelEncoder):
 class SalespersonEncoder(ModelEncoder):
     model = Salesperson
     properties = [
+        "id",
         "first_name",
         "last_name",
         "employee_id",
-        "id",
     ]
 
 
 class CustomerEncoder(ModelEncoder):
     model = Customer
     properties = [
+        "id",
         "first_name",
         "last_name",
         "address",
         "phone_number",
-        "id",
     ]
 
 
@@ -47,9 +47,9 @@ class SaleEncoder(ModelEncoder):
     ]
 
     encoders = {
-        "automobile": AutomobileVOEncoder(),
-        "salesperson": SalespersonEncoder(),
-        "customer": CustomerEncoder(),
+        "automobile": AutomobileVOEncoder,
+        "salesperson": SalespersonEncoder,
+        "customer": CustomerEncoder,
     }
 
 
@@ -61,7 +61,7 @@ def api_salespeople(request):
             {"salespeople": salespeople},
             encoder=SalespersonEncoder,
         )
-    else:   # POST
+    else:
         try:
             content = json.loads(request.body)
             salesperson = Salesperson.objects.create(**content)
@@ -131,49 +131,35 @@ def api_sales(request):
         return JsonResponse(
             {"sales": sales},
             encoder=SaleEncoder,
-            safe=False,
         )
-    else:
+    else:  # POST
         content = json.loads(request.body)
         try:
-            print("content from json.loads", content)
             vin = content["automobile"]
             automobile = AutomobileVO.objects.get(vin=vin)
-            print("automobile retrieved as matching vin:", automobile)
-            if automobile.sold:
-                return JsonResponse(
-                    {"message": "This automobile has been sold already!"}
-                )
-            else:
-                automobile.sold = True
-                automobile.save()
-                content["automobile"] = automobile
-        except AutomobileVO.DoesNotExist:
-            return JsonResponse(
-                {"message": "Invalid vin number"},
-                status=400,
-            )
-
-        try:
-            salesperson_id = content["salesperson"]
-            salesperson = Salesperson.objects.get(id=salesperson_id)
+            content["automobile"] = automobile
+            employee_id = content["salesperson"]
+            salesperson = Salesperson.objects.get(id=employee_id)
             content["salesperson"] = salesperson
-        except Salesperson.DoesNotExist:
-            return JsonResponse(
-                {"message": "Invalid salesperson id"},
-                status=400,
-            )
-
-        try:
             customer_id = content["customer"]
             customer = Customer.objects.get(id=customer_id)
             content["customer"] = customer
+
+        except AutomobileVO.DoesNotExist:
+            return JsonResponse(
+                {"message": "Automobile does not exist "},
+                status=400
+            )
+        except Salesperson.DoesNotExist:
+            return JsonResponse(
+                {"message": "Salesperson does not exist "},
+                status=400
+            )
         except Customer.DoesNotExist:
             return JsonResponse(
-                {"message": "Invalid customer id"},
-                status=400,
+                {"message": "Salesperson does not exist "},
+                status=400
             )
-
         sale = Sale.objects.create(**content)
         return JsonResponse(
             sale,
@@ -181,8 +167,7 @@ def api_sales(request):
             safe=False,
         )
 
-
-@require_http_methods(["DELETE", "GET"])
+@require_http_methods(["DELETE", "GET", "PUT"])
 def api_sale(request, pk):
     if request.method == "GET":
         try:
@@ -196,18 +181,34 @@ def api_sale(request, pk):
             response = JsonResponse({"message": "Does not exist"})
             response.status_code = 404
             return response
-    else:
+    elif request.method == "DELETE":
         try:
             sale = Sale.objects.get(id=pk)
-            automobile = AutomobileVO.objects.get(vin=sale.automobile.vin)
-            automobile.sold = False
-            automobile.save()
             sale.delete()
             return JsonResponse(
-                {"message": "deleted"}
+                sale,
+                encoder=SaleEncoder,
+                safe=False,
             )
         except Sale.DoesNotExist:
-            response = JsonResponse({"message": "Invalid Sale ID"})
+            return JsonResponse({"message": "Does not exist"})
+    else:  # PUT
+        try:
+            content = json.loads(request.body)
+            sale = Sale.objects.get(id=pk)
+
+            props = ["name"]
+            for prop in props:
+                if prop in content:
+                    setattr(sale, prop, content[prop])
+            sale.save()
+            return JsonResponse(
+                sale,
+                encoder=SaleEncoder,
+                safe=False,
+            )
+        except Sale.DoesNotExist:
+            response = JsonResponse({"message": "Does not exist"})
             response.status_code = 404
             return response
 
@@ -229,7 +230,7 @@ def api_customers(request):
                 encoder=CustomerEncoder,
                 safe=False,
             )
-        except ValueError:
+        except:
             response = JsonResponse(
                 {"message": "Could not create the customer"}
             )
@@ -281,13 +282,3 @@ def api_customer(request, pk):
             response = JsonResponse({"message": "Does not exist"})
             response.status_code = 404
             return response
-
-
-@require_http_methods(["GET"])
-def api_list_automobiles(request):
-    automobiles = AutomobileVO.objects.filter(sold=False)
-    return JsonResponse(
-        {"available_autos": automobiles},
-        encoder=AutomobileVOEncoder,
-        safe=False
-    )
